Add 'paused' label to 'prefect_info_deployments' metric (#47)
* Add 'paused' label to 'prefect_info_deployments' metric

- Adds the 'paused' label to the 'prefect_info_deployments' metric
- Also uses this value for the 'is_schedule_active' label, which is
  deprecated and always returns Null.
- Updates the docker-compose testing setup to use Prefect 3

Related to PLA-224

* Negate the 'paused' value for 'is_schedule_active'

'is_schedule_active' is the opposite of 'paused', so if we're using the
'paused' value we need to negate it.

* Fix null case

* Fix logic for negating value

If we get "null" back, we don't want to `not` that because we'll always
get False instead of "null". This separates the logic and only negates
the value if we don't get "null" back.